+++ hobj/__init__.py
@@ -1,9 +1,9 @@
 from hobj.benchmarks.mut_oneshot_benchmark import MutatorOneshotBenchmark
 from hobj.benchmarks.mut_highvar_benchmark import MutatorHighVarBenchmark
 
-from hobj.data_loaders.behavior import load_highvar_behavior, load_oneshot_behavior
+from hobj.data.behavior import load_highvar_behavior, load_oneshot_behavior
 
-from hobj.data_loaders.images import (
+from hobj.data.images import (
     load_image,
     list_image_ids,
     get_image_path,
@@ -13,7 +13,23 @@ from hobj.data_loaders.images import (
     load_imageset_meta_catch,
 )
 
+from hobj.learning_models.make_model import (
+    create_linear_learner,
+)
+
+from hobj.learning_models.random_guesser import RandomGuesser
+
 __all__ = [
+    # Image loader:
+    "list_image_ids",
+    "get_image_path",
+    "load_image",
+    # Learning Models:
+    "create_linear_learner",
+    "RandomGuesser",
+    # Benchmarks:
+    "MutatorHighVarBenchmark",
+    "MutatorOneshotBenchmark",
     # Raw behavior loaders
     "load_highvar_behavior",
     "load_oneshot_behavior",
@@ -22,11 +38,4 @@ __all__ = [
     "load_imageset_meta_oneshot",
     "load_imageset_meta_warmup",
     "load_imageset_meta_catch",
-    # Image loader:
-    "list_image_ids",
-    "load_image",
-    "get_image_path",
-    # Benchmarks:
-    "MutatorHighVarBenchmark",
-    "MutatorOneshotBenchmark",
 ]
