+++ tests/test_simulate_subtask.py
@@ -3,7 +3,7 @@ from typing import List
 import numpy as np
 
 from hobj.benchmarks.binary_classification.simulation import BinaryClassificationSubtask
-from hobj.learning_models import RandomGuesser
+from hobj.learning_models.random_guesser import RandomGuesser
 from hobj.types import ImageId
